+++ teacher/teacher_Grades.php
@@ -363,6 +363,8 @@ $page_title = 'Grade Entry';
 </div>
 
 <script src="<?php echo appAssetPath('vendor/bootstrap/bootstrap.bundle.min.js'); ?>"></script>
+<script src="<?php echo appAssetPath('js/offlineStorage.js'); ?>"></script>
+<script src="<?php echo appAssetPath('js/networkSync.js'); ?>"></script>
 <script src="<?php echo appAssetPath('js/main.js'); ?>"></script>
 <script>
 let gradingWeights = {